requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
lxml==4.9.3
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
pyaml==23.7.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
wxPython==4.2.1
